modify and move calculateTotal() to utils.dart
