+++ lib/cubit/page_cubit.dart
@@ -7,6 +7,7 @@ import 'package:onyx/store/page_store.dart';
 import 'package:equatable/equatable.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:nanoid/nanoid.dart';
+import 'package:onyx/utils/utils.dart';
 import 'package:replay_bloc/replay_bloc.dart';
 
 class PageState extends Equatable {
@@ -98,83 +99,6 @@ class PageCubit extends ReplayCubit<PageState> {
     emit(newPage);
   }
 
-  num calculateUntil(List<ListItemState> items, int untilIndex) {
-    final limit = untilIndex < items.length ? untilIndex : items.length;
-    if (limit == 0) return 0;
-
-    // Filter out items that have neither a number nor a valid operator
-    var working = List.of(items.take(limit))
-        .where((item) =>
-            item.number != null ||
-            (item.operator == Operator.add || item.operator == Operator.subtract || item.operator == Operator.multiply || item.operator == Operator.divide))
-        .toList();
-
-    // Reduce groups of 2+ adjacent items with the same (max) indent
-    while (true) {
-      if (working.length < 2) break;
-      final maxIndent = working.map((e) => e.indent).fold(0, (a, b) => a > b ? a : b);
-
-      // Find the first group of 2+ adjacent items with maxIndent
-      int start = -1, end = -1;
-      for (int i = 0; i < working.length;) {
-        if (working[i].indent == maxIndent) {
-          int j = i;
-          while (j < working.length && working[j].indent == maxIndent) j++;
-          if (j - i > 1) {
-            start = i;
-            end = j;
-            break;
-          }
-          i = j;
-        } else {
-          i++;
-        }
-      }
-      if (start == -1 || end == -1 || maxIndent == 0) break;
-
-      // Reduce the group left-to-right
-      num acc = working[start].number ?? 0;
-      for (int i = start + 1; i < end; i++) {
-        final op = working[i].operator;
-        final n = working[i].number ?? 0;
-        acc = switch (op) {
-          Operator.add => acc + n,
-          Operator.subtract => acc - n,
-          Operator.multiply => acc * n,
-          Operator.divide => n != 0 ? acc / n : 0,
-          _ => n,
-        };
-      }
-
-      // Replace the group with a single item (copy first, update number and indent)
-      final itemToCopy = working[start];
-      working.replaceRange(start, end, [itemToCopy.copyWith(number: acc, indent: maxIndent - 1)]);
-    }
-
-    // Sum up the remaining items
-    num sum = 0;
-    for (var item in working) {
-      final n = item.number ?? (item.operator == Operator.multiply || item.operator == Operator.divide ? 1 : 0);
-      switch (item.operator) {
-        case Operator.add:
-          sum += n;
-          break;
-        case Operator.subtract:
-          sum -= n;
-          break;
-        case Operator.multiply:
-          sum *= n;
-          break;
-        case Operator.divide:
-          sum /= n;
-          break;
-        default:
-          break;
-      }
-    }
-    return sum;
-  }
-
   void update(int index, ListItemState model) {
     final items = List.of(state.items, growable: true);
     items[index] = Parser.parse(model);
@@ -182,7 +106,7 @@ class PageCubit extends ReplayCubit<PageState> {
       PageState(
         items: items,
         index: state.index,
-        sum: calculateUntil(items, items.length),
+        sum: calculateTotal(items, items.length),
         created: state.created,
         modified: state.modified,
         title: state.title,
@@ -208,7 +132,7 @@ class PageCubit extends ReplayCubit<PageState> {
       PageState(
         items: items,
         index: index > 0 ? index - 1 : 0,
-        sum: calculateUntil(items, items.length),
+        sum: calculateTotal(items, items.length),
         created: state.created,
         modified: state.modified,
         title: state.title,
@@ -244,7 +168,7 @@ class PageCubit extends ReplayCubit<PageState> {
       PageState(
         items: items,
         index: items.length == 1 ? 0 : state.index + 1,
-        sum: calculateUntil(items, items.length),
+        sum: calculateTotal(items, items.length),
         created: state.created,
         modified: state.modified,
         title: state.title,
@@ -301,7 +225,7 @@ class PageCubit extends ReplayCubit<PageState> {
       PageState(
         items: items.mapIndexed((i, e) => e.copyWith(index: i)).toList(),
         index: items.indexOf(item),
-        sum: calculateUntil(items, items.length),
+        sum: calculateTotal(items, items.length),
         created: state.created,
         modified: state.modified,
         title: state.title,
